fix(pskreporter): round fractional decode periods
Preserve the final FT4 decode period when its fractional duration crosses UTC midnight. Keep the rounding in the shared decoded-time helper so standard and wideband spots use the same date rule, and cover the 7.5-second boundary.

+++ Network/DecodedTime.hpp
@@ -7,7 +7,7 @@
 namespace DecodedTime
 {
   QDateTime spotTime (QString const& encodedTime, QDateTime const& nowUtc,
-                      int periodSeconds);
+                      double periodSeconds);
 }
 
 #endif
